% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDHI.r
\name{CalculateDHI}
\alias{CalculateDHI}
\title{Calculate DHI value for drug MSI data}
\usage{
CalculateDHI(drug_csvfilepath,mask_csvfilepath,QuantLevel=32,Nu=1)
}
\arguments{
\item{drug_csvfilepath}{path of drug ion image file (csv format)}

\item{mask_csvfilepath}{path of tissue masked file (csv format)}

\item{QuantLevel}{Maximum possible gray-levels in drug ion image. default = 32}

\item{Nu}{Lowest size-zone value used for DHI calculation. default = 1}
}
\value{
DHI value for given drug ms file
}
\description{
This function calculates the DHI value for drug MSI data.
}
\details{
The algorithm is based on the following workflow:
\enumerate{
\item Derive new quantized drug image at user-defined qunatization value.
\item Derive gray-level size-zone matrix (GLSZM) for new drug quantized image (default value =32).
\item Derive homogeneous size-zone value from GLSZM at user-defined Nu value (default value =1).
\item Normalized overall value with complete tumor area which is obtained from tumor mask file.
}
}
\examples{
## load package
library(HomogenMSI)


## Input files path
drug_csvfilepath = "C:\\\\folder\\\\drug.csv"
mask_csvfilepath = "C:\\\\folder\\\\drug_masked.csv"

#Calculate DHI with default input parameters
print(CalculateDHI(drug_csvfilepath,mask_csvfilepath))

# calculate DHI with user-defined Qunantization level and Nu value
print(CalculateDHI(drug_csvfilepath,mask_csvfilepath,8,5))

}
\references{
\url{https://github.com/pietrofranceschi/HomogenMSI}
}
\seealso{
\code{\link{PreprocessingMSI}}
}
\author{
Mridula Prasad \email{mridula.prasad@fmach.it}
}
