% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHIimages.r
\name{DHIimages}
\alias{DHIimages}
\title{Set of synthetic images for homogeneity assessment}
\format{DHIimages is a list of 20 synthetic images showing different degrees of homogeneity. 5 basic images which simulate
real life experimental distributions are modified in the following four ways.
\enumerate{
\item In scenario 1, all images are with single gray level with different number of dark zones.
\item In scenarios 2 and 3, similar images contain 2 and 4 gray levels, respectively.
\item In scenario 4, all images are similar to the ones in scenario 3 with added salt-pepper noise.
}}
\usage{
data(DHIimages)
}
\description{
This dataset set contains several synthetic images for drug homogeneity assessment problem.
}
