% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingMSI.R
\name{PreprocessingMSI}
\alias{PreprocessingMSI}
\title{Pre-process MSI data}
\arguments{
\item{filepath}{path of MSI data folder in Analyze 7.5 format,}

\item{filename}{filename to save drug and corresponding mask image}

\item{mzs}{mass range for drug, tissue, internal standard (optional) ion images}
}
\value{
Returns csv files of drug ion and corresponding mask image for DHI calculation
}
\description{
Derive drug and mask ion images from MSI data.
}
\details{
\enumerate{ The function performs data pre-processing using follwoing steps:
\item Create mass bin of size 0.5
\item Extract drug, tissue and internal standard peaks for individual spectrum within their respective bins based on local maxima search.
\item Create tumor masked image based on tissue ion image.
\item Remove spatial noise using median filtering.
}
}
\examples{
## load library
library(HomogenMSI)

filepath  = 'C:\\\\folder\\\\drug' ## MSI data folder path

filename ='drug'  ## filename to save csv files
mzs = mzs = c(284.2, 281.2, 289.2)  ## mass list corresponds to the drug, tissue and internal standard mass

output = PreprocessingMSI(filepath,filename,mzs)
}
\references{
\url{https://github.com/pietrofranceschi/HomogenMSI}
}
\author{
Mridula Prasad \email{mridula.prasad@fmach.it}
}
