% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Calculate DHI value for artificially simulated images}
\usage{
testCalculateDHI(DrugFile,Nu=1)
}
\value{
DHI value for an input image
}
\description{
Examples to calculate DHI value over artifically simulated images.Tumor area for all calculation obatined from image A in case I. Images in all cases contain 3-5 gray-levels, hence input parameter number of quantization levels will skip here.
}
\examples{

# Image file for drug homogeneity index calculation
DHI =  testCalculateDHI(DHIimages[[5]])

 # setting Nu parameter value
 DHI =  testCalculateDHI(DHIimages[[5]],10)

}
