% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDHI.r
\name{testDHI}
\alias{testDHI}
\title{Perform a permutation test on DHI}
\usage{
testDHI(drugImg,permutations = 100, Nu=1, QuantLevel=NULL, maskImg=NULL)
}
\arguments{
\item{drugImg}{Input matrix containing the pixel intensities}

\item{permutations}{Number of permutations used to construct the null distribution. default = 100}

\item{Nu}{Lowest size-zone value used for DHI calculation. default = 1}

\item{QuantLevel}{Maximum possible gray-levels in drug ion image. default = 0, i.e., original image}

\item{maskImg}{Matrix identifying the pixels belonging to the tissue (1 tissue, 0 background). It must have the same size as drugImg}
}
\value{
A list containing the DHI calculated for the original and the permuted images
}
\description{
This function perform a permutation testing to assess the significance of the Drug Homogeneity Index.
}
\examples{
## load package
library(HomogenMSI)
data("DHIimages")

## Input drug and mask images

drugImg = DHIimages[[20]]
maskImg = DHIimages[[1]]
maskImg[maskImg !=0] =1

# test the significance of the DHI
output <- testDHI(drugImg,maskImg=maskImg)

# Plot the results
hist(output$DHIPermutedvalues, 
     xlim = range(c(output$DHIoriginalvalue,output$DHIPermutedvalues)), 
     col = "steelblue", 
     xlab = "DHI",
     main = "DHI Permutation test")
abline(v = output$DHIoriginalvalue, lty = 2)
}
\references{
\url{https://github.com/pietrofranceschi/HomogenMSI}
}
\author{
Mridula Prasad \email{mridula.prasad@fmach.it}
}
