% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDHI.r
\name{CalculateDHI}
\alias{CalculateDHI}
\title{Calculate DHI value for drug MSI data}
\usage{
CalculateDHI(drugImg,QuantLevel=0,Nu=1,maskImg=NULL)
}
\arguments{
\item{drugImg}{Input matrix containing the pixel intensities}

\item{maskImg}{Matrix identifying the pixels belonging to the tissue (1 tissue, 0 background). It must have the same size as drugImg}

\item{QuantLevel}{Maximum possible gray-levels in drug ion image. default = 0, i.e., original image}

\item{Nu}{Lowest size-zone value used for DHI calculation. default = 1}
}
\value{
DHI value for given drug ms file
}
\description{
This function calculates the DHI value for the drug image extracted from MSI data.
}
\details{
The algorithm is based on the following workflow:
\enumerate{
\item Derive new quantized drug image at user-defined qunatization value.
\item Derive gray-level size-zone matrix (GLSZM) for new drug quantized image (default value =0).
\item Derive homogeneous size-zone value from GLSZM at user-defined Nu value (default value =1).
\item Normalized overall value with complete tumor area which is obtained from tumor mask file.
}
}
\examples{
## load package
library(HomogenMSI)
data("DHIimages")

## Input drug and mask images

drugImg = DHIimages[[2]]
maskImg = DHIimages[[1]]
maskImg[maskImg !=0] =1

#Calculate DHI with default input parameters
print(CalculateDHI(drugImg,maskImg))

# calculate DHI with user-defined Nu value
print(CalculateDHI(drugImg,maskImg,QuantLevel=0,Nu=5))

}
\references{
\url{https://github.com/pietrofranceschi/HomogenMSI}
}
\author{
Mridula Prasad \email{mridula.prasad@fmach.it}
}
