% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDHI.r
\name{CalculateDHI}
\alias{CalculateDHI}
\title{Calculate DHI value for drug MSI data}
\usage{
CalculateDHI(drugImg,Nu=1,QuantLevel=NULL,maskImg=NULL)
}
\arguments{
\item{drugImg}{Input matrix containing the pixel intensities}

\item{maskImg}{Matrix identifying the pixels belonging to the tissue (1 tissue, 0 background). It must have the same size as drugImg}

\item{QuantLevel}{Maximum possible gray-levels in drug ion image}

\item{Nu}{Lowest size-zone value used for DHI calculation. Default = 1}
}
\value{
DHI value for the drug image
}
\description{
This function calculates the DHI value for the drug image extracted from MSI data.
}
\details{
The algorithm is based on the following workflow:
\enumerate{
\item Bin the drug image at user-defined qunatization value (optional).
\item Identify the pixel belonging to the tissue (optional)
\item Calculate the gray-level size-zone matrix (GLSZM).
\item Estimate the DHI
}
}
\examples{
## load package
library(HomogenMSI)
data("DHIimages")

## Input drug and mask images

drugImg = DHIimages[[2]]
maskImg = DHIimages[[1]]
maskImg[maskImg !=0] =1

#Calculate DHI with default input parameters
print(CalculateDHI(drugImg,maskImg=maskImg))

# calculate DHI with user-defined Nu value
print(CalculateDHI(drugImg,maskImg=maskImg,Nu=5))

}
\references{
\url{https://github.com/pietrofranceschi/HomogenMSI}
}
\author{
Mridula Prasad \email{mridula.prasad@fmach.it}
}
